feat: Embed additional text within generated workflows using globals.txt
This commit enhances the globals.txt configuration file, used by `make.sh`,
to include embedded text that will be incorporated into the final generated
workflows. By adding variables such as "{#GGUF_REQUIREMENTS}", the information
is easily editable in a single location. When executed with `make.sh`, this
text is seamlessly integrated into the workflow, providing users with
documentation and instructions.